// src/app/modules/common-registration/models/registration-tid.enum.ts
export enum RegistrationTid {
  GeneralObservation = 10,
  Incident = 11,
  Picture = 12,
  DangerObs = 13,
  DamageObs = 14,
  WeatherObservation = 21,
  SnowSurfaceObservation = 22,
  SnowCoverObs = 23,
  AvalancheDangerObs = 24,
  CompressionTest = 25,
  AvalancheObs = 26,
  AvalancheActivityObs = 27,
  AvalancheEvaluation = 28,
  AvalancheEvaluation2 = 30,
  AvalancheEvaluation3 = 31,
  AvalancheEvalProblem2 = 32,
  AvalancheActivityObs2 = 33,
  SnowTempObs = 34,
  DensityProfile = 35,
  SnowProfile2 = 36,
  StratProfile = 37,
  IceThickness = 50,
  IceCoverObs = 51,
  WaterLevel = 61,
  WaterLevel2 = 62,
  LandSlideObs = 71,
}